update: メッセージの日本語化。だいたい終了。

+++ source/app.d
@@ -735,8 +735,8 @@ void setRndSeed()
             seed ~= c;
 
     ulong seed1 , seed2;
-    seed1 = to!uint( seed[ 0 .. seed.length / 2 ] / 10 );
-    seed2 = to!uint( seed[ seed.length / 2 .. seed.length ] / 10 );
+    seed1 = to!ulong( seed[ 0 .. seed.length / 2 ] ) / 10;
+    seed2 = to!ulong( seed[ seed.length / 2 .. seed.length ] ) / 10;
 
     assert( seed1 + seed2 < int.max , "setRndSeed Error." );
